Minor update to bindings 2

+++ include/pocketpy/compiler.h
@@ -31,7 +31,7 @@ class Compiler {
     const Token& curr() const{ return tokens.at(i); }
     const Token& next() const{ return tokens.at(i+1); }
     const Token& err() const{
-        if(i >= static_cast<int>(tokens.size())) return prev();
+        if(static_cast<int>(i) >= static_cast<int>(tokens.size())) return prev();
         return curr();
     }
     void advance(int delta=1) { i += delta; }
